hf_space_deploy/requirements.txt
torch==2.1.0
torchvision==0.16.0
numpy==1.26.4
Pillow>=10.0.0
gradio==5.23.3
